pyproject.toml
[project]
name = "cuda"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "nvidia-cutlass-dsl>=4.2.1",
    "pandas>=2.3.3",
    "torch>=2.9.0",
    "triton>=3.5.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux'" }
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.pyright]
typeCheckingMode = "basic"
